Refactor editor components to implement WindowOverlay and streamline dialog handling
- Removed the DialogOverlay component and replaced it with the new WindowOverlay for improved dialog and palette management.
- Updated VMPanelDialog and VMImageUploadDialog to utilize WindowOverlay, enhancing the drag and resize functionality.
- Introduced useDraggableWindow hook to manage window positioning and resizing, promoting code reuse across dialog components.
- Created VMPanelPalette and VMHari components to support non-modal panel functionality, improving user interaction with the editor.

These changes enhance the editor's component architecture, providing a more flexible and intuitive user experience.
