+++ packages/editor/app/constants.ts
@@ -1,5 +1,3 @@
-export const PANEL_MIN_WIDTH = 300
-export const PANEL_MIN_HEIGHT = 300
 export const PANEL_INITIAL_WIDTH = 520
 export const PANEL_INITIAL_HEIGHT = 520
